Add a new extension event: TRANSLATED_PATH.
This event is useful to substitute the translated path.  Note I will not
promote this event in the generic version of PRoot because path
substitution should be done using the binding mechanism.  This will come
at no cost (and thus link2symlink extension should switch) once the VFS
is implemented in PRoot.